requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
flask-sqlalchemy==3.0.5
flask-login==0.6.2
flask-migrate==4.0.4
flask-wtf==1.1.1
email-validator==2.0.0